removed state from note-item, maintain editing state on each note object, add clear form method on form submission
